fix: refactor code and use seperated files for html components
