+++ op_tcg/frontend/html/modal.html
@@ -1,0 +1,5 @@
+<!-- The Modal -->
+<div id="myModal" class="modal">
+  <span class="modal-close" onclick="closeModal()">&times;</span>
+  <img class="modal-image" id="img01">
+</div>
