+++ op_tcg/frontend/styles/list_view.css
@@ -1,0 +1,86 @@
+/* Define color variables for light and dark themes */
+:root {
+  --background-color-light: #f9f9f9;
+  --border-color-light: #ccc;
+  --text-color-light: #333;
+  --modal-bg-color-light: rgba(0, 0, 0, 0.7);
+
+  --background-color-dark: #1e1e1e;
+  --border-color-dark: #444;
+  --text-color-dark: #f1f1f1;
+  --modal-bg-color-dark: rgba(255, 255, 255, 0.7);
+}
+
+/* Apply light theme by default */
+body {
+  background-color: var(--background-color-light);
+  color: var(--text-color-light);
+}
+
+.list-view {
+  list-style: none;
+  padding: 0;
+  margin: 0;
+}
+
+.list-item {
+  display: flex;
+  align-items: center;
+  border-bottom: 1px solid var(--border-color-light);
+  padding: 10px;
+  background-color: var(--background-color-light);
+}
+
+.list-item:last-child {
+  border-bottom: none;
+}
+
+.item-image {
+  width: 100px;
+  height: 100px;
+  margin-right: 20px;
+  border-radius: 50%;
+  overflow: hidden;
+}
+
+.item-image img {
+  width: 100%;
+  height: auto;
+}
+
+.item-details {
+  flex-grow: 1;
+}
+
+.item-title {
+  font-size: 1.2em;
+  margin: 0 0 5px;
+}
+
+.item-facts {
+  margin: 0;
+  padding: 0;
+  list-style: none;
+}
+
+.item-facts li {
+  margin-bottom: 5px;
+  font-size: 0.9em;
+}
+
+/* Detect dark mode preference and apply dark theme */
+@media (prefers-color-scheme: dark) {
+  body {
+    background-color: var(--background-color-dark);
+    color: var(--text-color-dark);
+  }
+
+  .list-item {
+    border-bottom: 1px solid var(--border-color-dark);
+    background-color: var(--background-color-dark);
+  }
+
+  .modal {
+    background-color: var(--modal-bg-color-dark);
+  }
+}
